// client/src/app/car/car.model.ts
export enum CarStatus {
  STOPPED = 'STOPPED',
  TAKING_IMAGE = 'TAKING_IMAGE',
  IMAGE_TAKEN = 'IMAGE_TAKEN',
  UPLOADING_IMAGE = 'UPLOADING_IMAGE',
  IMAGE_UPLOADED = 'IMAGE_UPLOADED',
  ANALYSING_IMAGE = 'ANALYSING_IMAGE',
  TRAFFIC_LIGHT_DETECTED = 'TRAFFIC_LIGHT_DETECTED',
  TRAFFIC_LIGHT_NOT_PRESENT = 'TRAFFIC_LIGHT_NOT_PRESENT',
  MOVING_FORWARD = 'MOVING_FORWARD',
  MOVING_BACKWARD = 'MOVING_BACKWARD',
}

export interface EventModel {
  type: CarStatus;
  data: any;
}